+++ src/imio/smartweb/core/browser/ia/ia.py
@@ -3,8 +3,14 @@ from imio.smartweb.core.contents import IPages
 from imio.smartweb.core.contents import ISectionText
 from imio.smartweb.core.utils import get_categories
 
+import json
+
 
 class ProcessCategorizeContentView(BaseProcessCategorizeContentView):
+
+    def __call__(self):
+        self._get_all_text()
+        return super(ProcessCategorizeContentView, self).__call__()
 
     def _get_all_text(self):
         all_text = ""
@@ -12,6 +18,12 @@ class ProcessCategorizeContentView(BaseProcessCategorizeContentView):
             for _id, obj in self.context.objectItems():
                 if ISectionText.providedBy(obj):
                     all_text += " " + getattr(obj.text, "output", "")
+        else:
+            # We don't have context so we're in ++add++
+            # Need to get page title on request
+            raw = self.request.get("BODY")
+            data = json.loads(raw)
+            all_text = data.get("formdata").get("form-widgets-IBasic-title", "")
         return all_text.strip()
 
     def _process_specific(self, all_text, results):
@@ -22,7 +34,8 @@ class ProcessCategorizeContentView(BaseProcessCategorizeContentView):
 
     def _process_category(self, all_text, results):
         # If agent doesn't register prevsouly a categorization so IA can bring categorization
-        if self.context.taxonomy_page_category is None:
+        # ++add++ context is Plone and hasn't got taxonomy_page_category property
+        if getattr(self.context, "taxonomy_page_category", None) is None:
             categories_taxo = get_categories()
             categories_voca = categories_taxo.makeVocabulary(self.current_lang).inv_data
             categories_voc = [